Refactor chat and database API endpoints; add filter and database retrieval functions
